Integrate WebAssembly demo package
Add support for the wasm-demo WebAssembly demo package by referencing
it in QEMU and VisionFive2 configurations.

The package is provided as a Buildroot overlay and includes install
rules to copy the relevant files to the root user's home directory on
the target system.

Additionally, patch a few wasm-demo dependencies to use versions built
by Buildroot, ensuring compatibility with the target environment.

Signed-off-by: Markku Kylänpää <markku.kylanpaa@vtt.fi>