Add WebAssembly example using WAT format
Introduce a simple WebAssembly example written in text notation (WAT)
that implements a basic integer calculator supporting addition,
subtraction, multiplication, and division.

This example serves as a test case for the WebAssembly bytecode
interpreter integrated into Keystone via the wasm-demo package.
The example is pakaged as wasm-examples package that is provided as7
a Buildroot overlay.

Signed-off-by: Markku Kylänpää <markku.kylanpaa@vtt.fi>